requirements.txt
langgraph>=0.2
langchain-openai>=0.1.0
langfuse>=2.0.0
qdrant-client>=1.9.0
langchain-qdrant>=0.1.0
paramiko>=3.4.0
fastapi>=0.110.0
uvicorn>=0.29.0
python-dotenv>=1.0.0
pydantic>=2.6.0
rich>=13.7.0
pydantic-settings>=2.0.0
pyyaml>=6.0
